feat: create a funciton for encodedSalt.
